requirements.txt
alembic==1.9.3
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector==2.2.9
SQLAlchemy==2.0.3
typing-extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
